requirements.txt
streamlit==1.31.1
langchain==0.1.6
langchain-google-genai==0.0.8
langchain-community==0.0.19
google-generativeai==0.3.2
google-search-results==2.4.2